move Axi4Adapter into lib.
